Basic Implementation based off of simple csv data for 2019 year
I wrote a quick program that will read off of the csv's that I downloaded from https://www.sports-reference.com/cbb/seasons/2019-school-stats.html.
It just puts the data into a dataframe with the appropriate headers, so that we can start doing analysis on the data itself. One thing to note and probably chnage is that the program deletes the headers from the txt file to load into the dataframe.